Add reset to branch prediction RAM
If a program ended with a while(true) loop, resetting the core failed, because the branch prediction ram still returned a tag match and the predicted jump address (leading to the while(true) loop again) is still used after reset. Invalidating the whole BRAM is not possible, but it is sufficient to return an invalid entry (all zeroes) with the reset.